updated and split up the individual functions
